segmentation/CBIM-Medical-Image-Segmentation/requirement.txt
absl-py
addict
astunparse
cachetools
charset-normalizer
clang
contextlib2
cycler
einops
flatbuffers
fvcore
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
imageio
iopath
keras
Keras-Preprocessing
kiwisolver
Markdown
matplotlib
#mkl-fft
#mkl-service
ml-collections
mmcv
monai
networkx
oauthlib
opencv-python
opt-einsum
packaging
pandas
Pillow
portalocker
protobuf
ptflops
pyasn1
pyasn1-modules
pyparsing
python-dateutil
pytz
PyWavelets
PyYAML
requests
requests-oauthlib
rsa
scikit-image
scipy
seaborn
SimpleITK
six
tabulate
tensorboard
#tensorboard-data-server
#tensorboard-plugin-wit
#tensorflow
#tensorflow-estimator
termcolor
thop
tifffile
timm
torch
torchaudio
torchvision
tqdm
typing-extensions
urllib3
Werkzeug
wrapt
yacs
yapf
numpy
